fix: 修复嵌入式下导出数据错误
